[c++] Changed APCAWindow to use double instead of floats

+++ cpp_project/include/coders/constant/apca/apca_window.h
@@ -9,8 +9,8 @@
 class APCAWindow: public Window {
 
 private:
-    int min;
-    int max;
+    double min;
+    double max;
 #if !MASK_MODE
     bool nan_window;
 #endif
@@ -21,7 +21,7 @@ public:
     bool isFull();
     bool isEmpty();
     void addFirstValue(std::string x);
-    bool updateConstantValue(int new_min, int new_max);
+    bool updateConstantValue(double new_min, double new_max);
 };
 
 #endif //CPP_PROJECT_APCA_WINDOW_H
